Use a single temporary file for Grammarly
Sadly, reloading only works for updating the same file, so having one for each
buffer does not work with the current solution.

There is still a problem with `open -a Grammarly` not doing anything if
Grammarly is already open, so the solution assumes that Grammarly is closed or
that the user is visiting the generated file in Grammarly.